add validation forms

+++ src/app/services/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import { HttpClient } from "@angular/common/http"; // Changez type import à regular import
 import { Injectable } from "@angular/core"
-import { BehaviorSubject, Observable, tap } from "rxjs"
+import {BehaviorSubject, catchError, Observable, switchMap, tap, throwError} from "rxjs"
+import Swal from "sweetalert2";
 
 @Injectable({
   providedIn: "root",
@@ -23,33 +24,50 @@ export class AuthService {
   }
 
 
+
   register(user: any): Observable<any> {
-    if (user.profilePicture === null) {
-      delete user.profilePicture;
-    }
-    return this.http.post(`${this.apiUrl}/users`, user).pipe(
-      tap((registeredUser) => {
-        this.http.put(`${this.apiUrl}/token/3ede`, { exist: true }).subscribe({
-          next: (tokenResponse) => {
-            console.log("Token mis à jour :", tokenResponse);
-          },
-          error: (error) => {
-            console.error("Erreur lors de la mise à jour du token", error);
+    return this.http.get(`${this.apiUrl}/users?email=${user.email}`).pipe(
+      switchMap((existingUser: any) => {
+        if (existingUser && existingUser.length > 0) {
+          Swal.fire({
+            title: 'Erreur',
+            text: 'L\'email est déjà utilisé. Veuillez en choisir un autre.',
+            icon: 'error',
+            confirmButtonText: 'OK'
+          });
+          return throwError(() => new Error('Email déjà utilisé.'));
+        } else {
+          if (user.profilePicture === null) {
+            delete user.profilePicture;
           }
-        });
-        const userData = {
-          email: user.email,
-          firstName: user.firstName,
-          lastName: user.lastName
-        };
+          return this.http.post(`${this.apiUrl}/users`, user).pipe(
+            tap((registeredUser) => {
+              this.http.put(`${this.apiUrl}/token/3ede`, { exist: true }).subscribe({
+                next: (tokenResponse) => {
+                  console.log("Token mis à jour :", tokenResponse);
+                },
+                error: (error) => {
+                  console.error("Erreur lors de la mise à jour du token", error);
+                }
+              });
+              const userData = {
+                email: user.email,
+                firstName: user.firstName,
+                lastName: user.lastName
+              };
+              console.log("Utilisateur enregistré :", userData);
+              localStorage.setItem("currentUser", JSON.stringify(userData));
+              console.log("hhhhhh" + localStorage.getItem("currentUser"));
 
-        console.log("Utilisateur enregistré :", userData);
-        localStorage.setItem("currentUser", JSON.stringify(userData));
-        console.log("hhhhhh"+localStorage.getItem("currentUser"));
-
-        this.currentUserSubject.next(userData);
-
-
+              this.currentUserSubject.next(userData);
+            })
+          );
+        }
+      }),
+      catchError((error) => {
+        console.error(error.message);
+        // Vous pouvez également renvoyer l'erreur au composant pour l'afficher
+        return throwError(() => error);
       })
     );
   }
@@ -90,6 +108,20 @@ export class AuthService {
     }
     return null;
   }
+  getUserFirst(): string | null {
+    const user = localStorage.getItem("currentUser");
+    if (user) {
+      try {
+        const parsedUser = JSON.parse(user);
+        return parsedUser.firstName || null;
+      } catch (error) {
+        console.error("Erreur lors de la récupération de l'email :", error);
+        return null;
+      }
+    }
+    return null;
+  }
+
